pyproject.toml
[project]
name = "newsletter-agent"
version = "0.1.0"
description = "Automated security & AI research intelligence digest"
readme = "README.md"
authors = [
    { name = "Deev Pal" }
]
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "anthropic>=0.100.0",
    "click>=8.1.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "resend>=2.0.0",
    "jinja2>=3.1.0",
    "python-dateutil>=2.9.0",
    "tavily-python>=0.5.0",
    "openai>=1.0.0",
    "firecrawl-py>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
newsletter = "newsletter_agent.cli:cli"

[build-system]
requires = ["uv_build>=0.11.14,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/newsletter_agent/delivery/templates.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
